mock-backend/backend-openapi-mongodb/requirements.txt
connexion
connexion[swagger-ui]
pymongo==3.12.1
psycopg2-binary
pytz==2021.1
pyopenssl
gunicorn
redis
jsonschema
requests>=2.22.0
flask_bcrypt
flask-cors
python-jose[cryptography]
twilio
phonenumbers
pycrypto
pika
better_profanity
Werkzeug~=2.0.0